requirements.txt
discord.py=1.*
requests=2.22.*
humanfriendly=4.18
